Rename the CalendarComponent to DatePicker
