+++ src/components/AssignmentPage/AssignmentPage.js
@@ -1,12 +1,13 @@
 import React, { useState } from "react";
-import CalendarComponent from "../CalendarComponent/CalendarComponent";
+import CalendarComponent from "../DatePicker/DatePicker";
 import { Container, Row, Col } from "react-bootstrap";
 import { withFirebase } from "utils/Firebase";
 import { compose } from "recompose";
 import "bootstrap/dist/css/bootstrap.min.css";
+import DatePicker from "../DatePicker/DatePicker";
 function AssignmentPage() {
   const [Date, setDate] = useState();
-  function handleCalendarChange(value) {
+  function handleDatePickerChange(value) {
     setDate(value);
   }
   return (
@@ -24,7 +25,7 @@ function AssignmentPage() {
             <div>Insert Student List component</div>
           </Col>
           <Col>
-            <CalendarComponent handleChange={handleCalendarChange} />
+            <DatePicker handleChange={handleDatePickerChange} />
           </Col>
         </Row>
       </Container>
