+++ src/components/DatePicker/DatePicker.js
@@ -1,8 +1,8 @@
 import React, { useState, setState, useEffect } from "react";
 import Calendar from "react-calendar/dist/entry.nostyle";
-import "CalendarComponent/CalendarComponent.css";
+import "DatePicker/DatePicker.css";
 import { withFirebase } from "utils/Firebase";
-function CalendarComponent(props) {
+function DatePicker(props) {
   function onChange(date) {
     props.handleChange({ date });
   }
@@ -21,4 +21,4 @@ function CalendarComponent(props) {
     </div>
   );
 }
-export default withFirebase(CalendarComponent);
+export default withFirebase(DatePicker);
